refactor: State 하위에 모든 핀이 쓰러진 상태, 핀이 남은 상태, 시작 상태 3가지로 나누어서 분리하고 하위에 클래스가 상속받도록 수정
