+++ src/main/java/bowling/domain/frame/exception/InvalidFrameRecordActionException.java
@@ -1,4 +1,4 @@
-package bowling.domain.frame;
+package bowling.domain.frame.exception;
 
 public class InvalidFrameRecordActionException extends RuntimeException {
     private static final String MESSAGE = "해당 프레임에 더이상 점수를 기록할 수 없습니다.";
