+++ src/main/java/bowling/domain/member/MemberName.java
@@ -1,5 +1,7 @@
 package bowling.domain.member;
 
+import bowling.domain.member.exception.InvalidMemberNameLengthException;
+
 public class MemberName {
     private static final int FIXED_NAME_LENGTH = 3;
     private final String name;
