+++ src/main/java/bowling/domain/member/exception/InvalidMemberNameLengthException.java
@@ -1,4 +1,4 @@
-package bowling.domain.member;
+package bowling.domain.member.exception;
 
 public class InvalidMemberNameLengthException extends RuntimeException {
     private static final String MESSAGE = "참여자 이름은 3글자여야 합니다.";
