+++ src/main/java/bowling/domain/score/exception/InvalidScoreException.java
@@ -1,4 +1,4 @@
-package bowling.domain.score;
+package bowling.domain.score.exception;
 
 public class InvalidScoreException extends RuntimeException {
     private static final String MESSAGE = "점수는 한번에 1에서 10사이여야 합니다.";
