+++ src/main/java/bowling/domain/state/FrameReady.java
@@ -3,6 +3,9 @@ package bowling.domain.state;
 import bowling.domain.pin.Pin;
 import bowling.domain.score.Score;
 import bowling.domain.score.Scores;
+import bowling.domain.state.all.Strike;
+import bowling.domain.state.rest.Gutter;
+import bowling.domain.state.rest.Ordinary;
 
 import static bowling.domain.frame.Frame.MAX_TRY_COUNT;
