+++ src/main/java/bowling/domain/state/all/Spare.java
@@ -1,0 +1,9 @@
+package bowling.domain.state.all;
+
+import bowling.domain.score.Scores;
+
+public class Spare extends AllPinsFell {
+    public Spare(int leftTry, Scores scores) {
+        super(leftTry, scores);
+    }
+}
