+++ src/main/java/bowling/domain/state/exception/InvalidLeftTryException.java
@@ -1,4 +1,4 @@
-package bowling.domain.state;
+package bowling.domain.state.exception;
 
 public class InvalidLeftTryException extends RuntimeException {
     private static final String MESSAGE = "더이상 시도할 수 없습니다.";
