+++ src/main/java/bowling/domain/state/exception/InvalidRecordInOrdinaryException.java
@@ -1,4 +1,4 @@
-package bowling.domain.state;
+package bowling.domain.state.exception;
 
 public class InvalidRecordInOrdinaryException extends RuntimeException {
     private static final String MESSAGE = "해당 핀은 입력이 불가능합니다. 현재 입력된 핀 수는 %d";
