+++ src/test/java/bowling/domain/score/ScoreTest.java
@@ -1,5 +1,6 @@
 package bowling.domain.score;
 
+import bowling.domain.score.exception.InvalidScoreException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
