+++ src/test/java/bowling/domain/state/GutterTest.java
@@ -3,6 +3,9 @@ package bowling.domain.state;
 import bowling.domain.pin.Pin;
 import bowling.domain.score.Score;
 import bowling.domain.score.Scores;
+import bowling.domain.state.all.Spare;
+import bowling.domain.state.rest.Gutter;
+import bowling.domain.state.rest.Ordinary;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
